feat: break cli working into dir
