+++ Notes.md
@@ -1,0 +1,3 @@
+Hey, this files will be for notes and anything we want to keep track of. 
+
+- For temp basis I have created a folder with name `data` to test things out.
