+++ data/Pradumna.json
@@ -1,0 +1,7 @@
+{
+  "name": "a",
+  "type": "Personal",
+  "bio": "asas",
+  "links": [],
+  "milestones": []
+}
