+++ src/addtestimonial/addTestimonial.js
@@ -1,0 +1,5 @@
+const updateJson = () => {
+    console.log("Adding a testimonial");
+}
+
+module.exports = updateJson;
