+++ src/createtestimonial/createTestimonial.js
@@ -1,0 +1,5 @@
+const createTestimonial = () => {
+    console.log("Creating a testimonial");
+}
+
+module.exports = createTestimonial;
