+++ src/updatejson/updateJson.js
@@ -1,0 +1,5 @@
+const updateJson = () => {
+    console.log("Updating an existing JSON file");
+}
+
+module.exports = updateJson;
